+++ packages/backend/src/models/entities/Role.ts
@@ -136,7 +136,7 @@ export class Role {
 	@Column('jsonb', {
 		default: { },
 	})
-	public options: Record<string, {
+	public policies: Record<string, {
 		useDefault: boolean;
 		priority: number;
 		value: any;